+++ demo/src/controller/HelloController.ts
@@ -3,26 +3,29 @@
  * @Usage: 接收处理路由参数
  * @Author: xxx
  * @Date: 2020-12-22 15:24:25
- * @LastEditTime: 2021-12-23 00:59:23
+ * @LastEditTime: 2024-01-04 07:45:42
  */
 
-import { KoattyContext, Controller, BaseController, Autowired, RequestMapping, RequestBody } from '../../../src/index';
+import {
+  KoattyContext, Controller, Autowired, RequestMapping,
+  RequestBody, IController
+} from '../../../src/index';
 import { App } from '../App';
 import { SayHelloRequestDto } from '../dto/SayHelloRequestDto';
 import { SayHelloReplyDto } from '../dto/SayHelloReplyDto';
 import { Validated } from 'koatty_validation';
 
 @Controller('/Hello') // Consistent with proto.service name
-export class HelloController extends BaseController {
+export class HelloController {
   app: App;
   ctx: KoattyContext;
 
   /**
-   * Custom constructor
+   * constructor
    *
    */
-  init() {
-    //todo
+  constructor(ctx: KoattyContext) {
+    this.ctx = ctx;
   }