+++ demo/src/service/TestService.ts
@@ -3,17 +3,17 @@
  * @Usage: 处理具体业务逻辑
  * @Author: xxx
  * @Date: 2020-12-22 15:59:51
- * @LastEditTime: 2021-12-23 01:03:04
+ * @LastEditTime: 2024-01-16 01:00:28
  */
 
-import { Service, BaseService, Autowired, Logger } from '../../../src/index';
+import { Service, Autowired, Logger } from '../../../src/index';
 import { App } from '../App';
 import { UserDto } from '../dto/UserDto';
 // import { Scheduled, SchedulerLock } from "koatty_schedule";
 // import { CacheAble, CacheEvict, GetCacheStore } from "koatty_cacheable";
 
 @Service()
-export class TestService extends BaseService {
+export class TestService {
   app: App;
 
   /**