+++ src/core/Trace.ts
@@ -3,14 +3,14 @@
  * @Usage: 
  * @Author: richen
  * @Date: 2023-12-14 22:29:34
- * @LastEditTime: 2023-12-21 20:14:56
+ * @LastEditTime: 2024-01-16 22:41:52
  * @License: BSD (3-Clause)
  * @Copyright (c): <richenlin(at)gmail.com>
  */
 import { Koatty } from "koatty_core";
 import { Trace, TraceOptions } from "koatty_trace";
 
-export async function TraceHandler(app: Koatty) {
+export function TraceHandler(app: Koatty) {
   const timeout = (app.config('http_timeout') || 10) * 1000;
   const encoding = app.config('encoding') || 'utf-8';
   const openTrace = app.config("open_trace") || false;
@@ -25,5 +25,5 @@ export async function TraceHandler(app: Koatty) {
     OpenTrace: openTrace,
     AsyncHooks: asyncHooks,
   }
-  app.use(await Trace(options, app));
+  app.use(Trace(options, app));
 }