+++ src/index.ts
@@ -3,18 +3,18 @@
  * @Usage: 
  * @Author: richen
  * @Date: 2023-12-09 21:56:32
- * @LastEditTime: 2023-12-20 20:28:04
+ * @LastEditTime: 2024-01-16 01:14:19
  * @License: BSD (3-Clause)
  * @Copyright (c): <richenlin(at)gmail.com>
  */
 
 export * from "./core/Bootstrap";
-export * from "./component/Components";
-export * from "./component/BaseService";
-export * from "./component/BaseController";
+export * from "./component/Component";
+export * from "./component/Output";
 export * from "koatty_core";
 export * from "koatty_container";
 export * from "koatty_exception";
+export * from "koatty_router";
 export * from "koatty_serve";
 export { Value, Config } from "koatty_config";
 export { Helper } from "./util/Helper";